000_shoptalk_contact_scraper/requirements.txt
pandas
pillow
pytesseract
numpy
textdistance[extras]